deprecated LearningRateLogger
