+++ tests/callbacks/test_lr_monitor.py
@@ -1,7 +1,7 @@
 import pytest
 
 from pytorch_lightning import Trainer
-from pytorch_lightning.callbacks import LearningRateMonitor
+from pytorch_lightning.callbacks import LearningRateLogger, LearningRateMonitor
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.base import EvalModelTemplate
 import tests.base.develop_utils as tutils
@@ -125,3 +125,11 @@ def test_lr_monitor_param_groups(tmpdir):
         'Number of learning rates logged does not match number of param groups'
     assert all([k in ['lr-Adam/pg1', 'lr-Adam/pg2'] for k in lr_monitor.lrs.keys()]), \
         'Names of learning rates not set correctly'
+
+
+def test_lr_logger_deprecated(tmpdir):
+    """ Test the deprecation warning for LearningRateLogger. """
+    tutils.reset_seed()
+
+    with pytest.warns(DeprecationWarning, match='is now `LearningRateMonitor`'):
+        lr_logger = LearningRateLogger()
